requirements.txt
pymupdf>=1.24.0
pytesseract>=0.3.13
opencv-python>=4.10.0
Pillow>=10.0.0
numpy>=1.26.0
pydantic>=2.7.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
anthropic>=0.40.0
python-dotenv>=1.0.0
